backend/auth-service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysqlclient
gunicorn
